+++ qiskit/providers/ibmq/ibmqbackendservice.py
@@ -17,6 +17,7 @@
 import logging
 from typing import Dict, List, Callable, Optional, Any, Union
 from types import SimpleNamespace
+from datetime import datetime
 
 from qiskit.providers import JobStatus, QiskitBackendNotFoundError  # type: ignore[attr-defined]
 from qiskit.providers.providerutils import filter_backends
@@ -99,6 +100,8 @@ class IBMQBackendService(SimpleNamespace):
             backend_name: Optional[str] = None,
             status: Optional[Union[JobStatus, str]] = None,
             job_name: Optional[str] = None,
+            start_datetime: Optional[datetime] = None,
+            end_datetime: Optional[datetime] = None,
             db_filter: Optional[Dict[str, Any]] = None
     ) -> List[IBMQJob]:
         """Return a list of jobs from the API.
@@ -122,6 +125,10 @@ class IBMQBackendService(SimpleNamespace):
                 and `regular expressions
                 <https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Regular_Expressions>
                 `_ can be used.
+            start_datetime: filter by start date. This is used to find jobs
+                whose creation dates are after (greater than or equal to) this date/time.
+            end_datetime: filter by end date. This is used to find jobs
+                whose creation dates are before (less than or equal to) this date/time.
             db_filter: `loopback-based filter
                 <https://loopback.io/doc/en/lb2/Querying-data.html>`_.
                 This is an interface to a database ``where`` filter. Some
@@ -138,12 +145,6 @@ class IBMQBackendService(SimpleNamespace):
                                  'qasms.result.data.counts.00': {'gt': 400},
                                  'qasms.result.data.counts.11': {'gt': 400}}
                   job_list = backend.jobs(limit=5, db_filter=cnts_filter)
-
-                Filter last five jobs from 30 days ago::
-
-                   past_date = datetime.datetime.now() - datetime.timedelta(days=30)
-                   date_filter = {'creationDate': {'lt': past_date.isoformat()}}
-                   job_list = backend.jobs(limit=5, db_filter=date_filter)
 
         Returns:
             list of IBMQJob instances
@@ -180,8 +181,17 @@ class IBMQBackendService(SimpleNamespace):
         if job_name:
             api_filter['name'] = {"regexp": job_name}
 
+        if start_datetime and end_datetime:
+            api_filter['creationDate'] = {
+                'between': [start_datetime.isoformat(), end_datetime.isoformat()]
+            }
+        elif start_datetime:
+            api_filter['creationDate'] = {'gte': start_datetime.isoformat()}
+        elif end_datetime:
+            api_filter['creationDate'] = {'lte': end_datetime.isoformat()}
+
         if db_filter:
-            # status takes precedence over db_filter for same keys
+            # Argument filters takes precedence over db_filter for same keys
             api_filter = {**db_filter, **api_filter}
 
         # Retrieve the requested number of jobs, using pagination. The API