+++ test/ibmq/test_ibmq_job.py
@@ -15,10 +15,8 @@
 """IBMQJob Test."""
 
 import time
-import warnings
 from concurrent import futures
-from datetime import datetime
-from unittest import skip
+from datetime import datetime, timedelta
 
 import numpy
 from scipy.stats import chi2_contingency
@@ -30,8 +28,7 @@ from qiskit.providers.ibmq.ibmqbackend import IBMQRetiredBackend
 from qiskit.providers.ibmq.exceptions import IBMQBackendError
 from qiskit.providers.ibmq.ibmqfactory import IBMQFactory
 from qiskit.providers.ibmq.job.ibmqjob import IBMQJob
-from qiskit.providers.ibmq.job.exceptions import (IBMQJobFailureError,
-                                                  IBMQJobInvalidStateError)
+from qiskit.providers.ibmq.job.exceptions import IBMQJobInvalidStateError
 from qiskit.test import slow_test
 from qiskit.compiler import assemble, transpile
 from qiskit.result import Result
@@ -216,36 +213,17 @@ class TestIBMQJob(JobTestCase):
         self.assertTrue(can_cancel)
         self.assertTrue(job.status() is JobStatus.CANCELLED)
 
-    @requires_device
-    def test_get_jobs_from_backend(self, backend):
-        """Test retrieving jobs from a backend."""
-        job_list = backend.jobs(limit=5, skip=0)
-        for job in job_list:
-            self.assertTrue(isinstance(job.job_id(), str))
-
-    @requires_device
     @requires_provider
-    def test_get_jobs_from_backend_service(self, backend, provider):
-        """Test retrieving jobs from backend service."""
+    def test_retrieve_jobs(self, provider):
+        """Test retrieving jobs."""
+        backend = provider.get_backend('ibmq_qasm_simulator')
         job_list = provider.backends.jobs(backend_name=backend.name(), limit=5, skip=0)
         for job in job_list:
             self.assertTrue(isinstance(job.job_id(), str))
 
     @requires_provider
-    def test_retrieve_job_backend(self, provider):
-        """Test retrieving a single job from a backend."""
-        backend = provider.get_backend('ibmq_qasm_simulator')
-        qobj = assemble(transpile(self._qc, backend=backend), backend=backend)
-        job = backend.run(qobj)
-
-        retrieved_job = backend.retrieve_job(job.job_id())
-        self.assertEqual(job.job_id(), retrieved_job.job_id())
-        self.assertEqual(job.result().get_counts(), retrieved_job.result().get_counts())
-        self.assertEqual(job.qobj().to_dict(), qobj.to_dict())
-
-    @requires_provider
-    def test_retrieve_job_backend_service(self, provider):
-        """Test retrieving a single job from backend service."""
+    def test_retrieve_job(self, provider):
+        """Test retrieving a single job."""
         backend = provider.get_backend('ibmq_qasm_simulator')
 
         qobj = assemble(transpile(self._qc, backend=backend), backend=backend)
@@ -288,103 +266,128 @@ class TestIBMQJob(JobTestCase):
                               real_backend.retrieve_job, job_sim.job_id())
         self.assertIn('belongs to', str(context_manager.warning))
 
-    @requires_device
-    def test_retrieve_job_error_backend(self, backend):
-        """Test retrieving an invalid job from a backend."""
-        self.assertRaises(IBMQBackendError, backend.retrieve_job, 'BAD_JOB_ID')
-
     @requires_provider
-    def test_retrieve_job_error_backend_service(self, provider):
-        """Test retrieving an invalid job from backend service."""
+    def test_retrieve_job_error(self, provider):
+        """Test retrieving an invalid job."""
         self.assertRaises(IBMQBackendError, provider.backends.retrieve_job, 'BAD_JOB_ID')
 
-    @requires_device
-    def test_get_jobs_filter_job_status_backend(self, backend):
-        """Test retrieving jobs from a backend filtered by status."""
-        job_list = backend.jobs(limit=5, skip=0, status=JobStatus.DONE)
-        for job in job_list:
-            self.assertTrue(job.status() is JobStatus.DONE)
-
-    @requires_device
     @requires_provider
-    def test_get_jobs_filter_job_status_backend_service(self, backend, provider):
-        """Test retrieving jobs from backend service filtered by status."""
+    def test_retrieve_jobs_status(self, provider):
+        """Test retrieving jobs filtered by status."""
+        backend = provider.get_backend('ibmq_qasm_simulator')
         job_list = provider.backends.jobs(backend_name=backend.name(),
                                           limit=5, skip=0, status=JobStatus.DONE)
+
+        self.assertTrue(job_list)
         for job in job_list:
             self.assertTrue(job.status() is JobStatus.DONE)
 
     @requires_provider
-    def test_get_jobs_filter_counts_backend(self, provider):
-        """Test retrieving jobs from a backend filtered by counts."""
-        # TODO: consider generalizing backend name
-        # TODO: this tests depends on the previous executions of the user
+    def test_retrieve_jobs_start_datetime(self, provider):
+        """Test retrieving jobs created after a specified datetime."""
         backend = provider.get_backend('ibmq_qasm_simulator')
-        my_filter = {'backend.name': 'ibmq_qasm_simulator',
-                     'shots': 1024,
-                     'qasms.result.data.counts.00': {'lt': 500}}
-        self.log.info('searching for at most 5 jobs with 1024 shots, a count '
-                      'for "00" of < 500, on the ibmq_qasm_simulator backend')
+        past_month = datetime.now() - timedelta(days=30)
+        past_month_str = past_month.strftime('%Y-%m-%dT%H:%M:%S.%fZ')
 
-        with warnings.catch_warnings():
-            # Disable warnings from pre-qobj jobs.
-            warnings.filterwarnings('ignore',
-                                    category=DeprecationWarning,
-                                    module='qiskit.providers.ibmq.ibmqbackend')
-            job_list = backend.jobs(limit=5, skip=0, db_filter=my_filter)
-
-        for i, job in enumerate(job_list):
-            self.log.info('match #%d', i)
-            result = job.result()
-            self.assertTrue(any(cresult.data.counts.to_dict()['0x0'] < 500
-                                for cresult in result.results))
-
-    @requires_provider
-    def test_get_jobs_filter_counts_backend_service(self, provider):
-        """Test retrieving jobs from backend service filtered by counts."""
-        # TODO: consider generalizing backend name
-        # TODO: this tests depends on the previous executions of the user
-        backend = provider.get_backend('ibmq_qasm_simulator')
-
-        my_filter = {'backend.name': 'ibmq_qasm_simulator',
-                     'shots': 1024,
-                     'qasms.result.data.counts.00': {'lt': 500}}
-        self.log.info('searching for at most 5 jobs with 1024 shots, a count '
-                      'for "00" of < 500, on the ibmq_qasm_simulator backend')
-
-        with warnings.catch_warnings():
-            # Disable warnings from pre-qobj jobs.
-            warnings.filterwarnings('ignore',
-                                    category=DeprecationWarning,
-                                    module='qiskit.providers.ibmq.ibmqbackend')
-            job_list = provider.backends.jobs(backend_name=backend.name(),
-                                              limit=5, skip=0, db_filter=my_filter)
-
-        for i, job in enumerate(job_list):
-            self.log.info('match #%d', i)
-            result = job.result()
-            self.assertTrue(any(cresult.data.counts.to_dict()['0x0'] < 500
-                                for cresult in result.results))
-
-    @requires_device
-    def test_get_jobs_filter_date_backend(self, backend):
-        """Test retrieving jobs from a backend filtered by date."""
-        date_today = datetime.now().isoformat()
-        my_filter = {'creationDate': {'lt': date_today}}
-        job_list = backend.jobs(limit=5, db_filter=my_filter)
-
+        job_list = provider.backends.jobs(backend_name=backend.name(),
+                                          limit=5, skip=0, start_datetime=past_month)
         self.assertTrue(job_list)
-        self.log.info('found %s matching jobs', len(job_list))
         for i, job in enumerate(job_list):
-            self.log.info('match #%d: %s', i, job.creation_date())
-            self.assertTrue(job.creation_date() < date_today)
+            self.assertTrue(job.creation_date() >= past_month_str,
+                            '{}) job creation_date {} is not '
+                            'greater than or equal to past month: {}'
+                            .format(i, job.creation_date(), past_month_str))
 
-    @requires_device
     @requires_provider
-    def test_get_jobs_filter_date_backend_service(self, backend, provider):
-        """Test retrieving jobs from backend service filtered by date."""
-        date_today = datetime.now().isoformat()
-        my_filter = {'creationDate': {'lt': date_today}}
+    def test_retrieve_jobs_end_datetime(self, provider):
+        """Test retrieving jobs created before a specified datetime."""
+        backend = provider.get_backend('ibmq_qasm_simulator')
+        past_month = datetime.now() - timedelta(days=30)
+        past_month_str = past_month.strftime('%Y-%m-%dT%H:%M:%S.%fZ')
+
+        job_list = provider.backends.jobs(backend_name=backend.name(),
+                                          limit=5, skip=0, end_datetime=past_month)
+        self.assertTrue(job_list)
+        for i, job in enumerate(job_list):
+            self.assertTrue(job.creation_date() <= past_month_str,
+                            '{}) job creation_date {} is not '
+                            'less than or equal to past month: {}'
+                            .format(i, job.creation_date(), past_month_str))
+
+    @requires_provider
+    def test_retrieve_jobs_between_datetimes(self, provider):
+        """Test retrieving jobs created between two specified datetimes."""
+        backend = provider.get_backend('ibmq_qasm_simulator')
+        date_today = datetime.now()
+
+        past_month = date_today - timedelta(30)
+        past_month_str = past_month.strftime('%Y-%m-%dT%H:%M:%S.%fZ')
+        past_two_month = date_today - timedelta(60)
+        past_two_month_str = past_two_month.strftime('%Y-%m-%dT%H:%M:%S.%fZ')
+
+        job_list = provider.backends.jobs(backend_name=backend.name(), limit=5, skip=0,
+                                          start_datetime=past_two_month, end_datetime=past_month)
+        self.assertTrue(job_list)
+        for i, job in enumerate(job_list):
+            self.assertTrue((past_two_month_str <= job.creation_date() <= past_month_str),
+                            '{}) job creation date {} is not '
+                            'between past two month {} and past month {}'
+                            .format(i, past_two_month_str, job.creation_date(), past_month_str))
+
+    @requires_provider
+    def test_retrieve_jobs_between_datetimes_not_overriden(self, provider):
+        """Test retrieving jobs created between two specified datetimes
+        and ensure `db_filter` does not override datetime arguments."""
+        backend = provider.get_backend('ibmq_qasm_simulator')
+        date_today = datetime.now()
+
+        past_two_month = date_today - timedelta(30)
+        past_two_month_str = past_two_month.strftime('%Y-%m-%dT%H:%M:%S.%fZ')
+        past_three_month = date_today - timedelta(60)
+        past_three_month_str = past_three_month.strftime('%Y-%m-%dT%H:%M:%S.%fZ')
+
+        # Used for `db_filter`, should not override `start_datetime` and `end_datetime` arguments.
+        past_ten_days = date_today - timedelta(10)
+
+        job_list = provider.backends.jobs(backend_name=backend.name(), limit=5, skip=0,
+                                          start_datetime=past_three_month,
+                                          end_datetime=past_two_month,
+                                          db_filter={'creationDate': {'gt': past_ten_days}})
+        self.assertTrue(job_list)
+        for i, job in enumerate(job_list):
+            self.assertTrue((past_three_month_str <= job.creation_date() <= past_two_month_str),
+                            '{}) job creation date {} is not '
+                            'between past three month {} and past two month {}'
+                            .format(i, past_three_month_str,
+                                    job.creation_date(), past_two_month_str))
+
+    @requires_provider
+    def test_retrieve_jobs_filter_counts(self, provider):
+        """Test retrieving jobs filtered by counts."""
+        # TODO: consider generalizing backend name
+        # TODO: this tests depends on the previous executions of the user
+        backend = provider.get_backend('ibmq_qasm_simulator')
+
+        my_filter = {'backend.name': 'ibmq_qasm_simulator',
+                     'shots': 1024,
+                     'qasms.result.data.counts.00': {'lt': 500}}
+
+        job_list = provider.backends.jobs(backend_name=backend.name(),
+                                          limit=5, skip=0, db_filter=my_filter)
+
+        for job in job_list:
+            result = job.result()
+            self.assertTrue(any(cresult.data.counts.to_dict()['0x0'] < 500
+                                for cresult in result.results))
+
+    @requires_provider
+    def test_retrieve_jobs_filter_date(self, provider):
+        """Test retrieving jobs filtered by date."""
+        backend = provider.get_backend('ibmq_qasm_simulator')
+        date_today = datetime.now()
+        date_today_str = date_today.strftime('%Y-%m-%dT%H:%M:%S.%fZ')
+
+        my_filter = {'creationDate': {'lt': date_today.isoformat()}}
         job_list = provider.backends.jobs(backend_name=backend.name(),
                                           limit=5, db_filter=my_filter)
 
@@ -392,7 +395,9 @@ class TestIBMQJob(JobTestCase):
         self.log.info('found %s matching jobs', len(job_list))
         for i, job in enumerate(job_list):
             self.log.info('match #%d: %s', i, job.creation_date())
-            self.assertTrue(job.creation_date() < date_today)
+            self.assertTrue(job.creation_date() < date_today_str,
+                            '{}) job.creation_date: {}, date_today: {}'
+                            .format(i, job.creation_date(), date_today_str))
 
     @requires_provider
     def test_double_submit_fails(self, provider):
@@ -405,38 +410,6 @@ class TestIBMQJob(JobTestCase):
         with self.assertRaises(IBMQJobInvalidStateError):
             job.submit()
 
-    @requires_provider
-    def test_retrieve_failed_job_simulator(self, provider):
-        """Test retrieving job error messages from a simulator backend."""
-        backend = provider.get_backend('ibmq_qasm_simulator')
-
-        qc_new = transpile(self._qc, backend)
-        qobj = assemble([qc_new, qc_new], backend=backend)
-        qobj.experiments[1].instructions[1].name = 'bad_instruction'
-
-        job = backend.run(qobj)
-        with self.assertRaises(IBMQJobFailureError):
-            job.result()
-
-        new_job = provider.backends.retrieve_job(job.job_id())
-        message = new_job.error_message()
-        self.assertIn('Experiment 1: ERROR', message)
-
-    @run_on_device
-    def test_retrieve_failed_job_device(self, provider, backend):
-        """Test retrieving a failed job from a device backend."""
-        qc_new = transpile(self._qc, backend)
-        qobj = assemble([qc_new, qc_new], backend=backend)
-        qobj.experiments[1].instructions[1].name = 'bad_instruction'
-
-        job = backend.run(qobj)
-        with self.assertRaises(IBMQJobFailureError):
-            job.result(timeout=180)
-
-        new_job = provider.backends.retrieve_job(job.job_id())
-        self.assertTrue(new_job.error_message())
-
-    @skip('Remove skip once simulator returns schema complaint partial results.')
     @requires_provider
     def test_retrieve_failed_job_simulator_partial(self, provider):
         """Test retrieving partial results from a simulator backend."""