+++ test/ibmq/test_ibmq_job_attributes.py
@@ -16,6 +16,7 @@
 
 import time
 from unittest import mock
+import re
 
 from qiskit import ClassicalRegister, QuantumCircuit, QuantumRegister
 from qiskit.providers import JobStatus
@@ -62,33 +63,8 @@ class TestIBMQJobAttributes(JobTestCase):
         _ = job.properties()
 
     @requires_provider
-    def test_job_name_backend(self, provider):
-        """Test using job names on a backend."""
-        backend = provider.get_backend('ibmq_qasm_simulator')
-        qobj = assemble(transpile(self._qc, backend=backend), backend=backend)
-
-        # Use a unique job name
-        job_name = str(time.time()).replace('.', '')
-        job_id = backend.run(qobj, job_name=job_name).job_id()
-        job = backend.retrieve_job(job_id)
-        self.assertEqual(job.name(), job_name)
-
-        # Check using partial matching.
-        job_name_partial = job_name[8:]
-        retrieved_jobs = backend.jobs(job_name=job_name_partial)
-        self.assertGreaterEqual(len(retrieved_jobs), 1)
-        retrieved_job_ids = {job.job_id() for job in retrieved_jobs}
-        self.assertIn(job_id, retrieved_job_ids)
-
-        # Check using regular expressions.
-        job_name_regex = '^{}$'.format(job_name)
-        retrieved_jobs = backend.jobs(job_name=job_name_regex)
-        self.assertEqual(len(retrieved_jobs), 1)
-        self.assertEqual(job_id, retrieved_jobs[0].job_id())
-
-    @requires_provider
-    def test_job_name_backend_service(self, provider):
-        """Test using job names on backend service."""
+    def test_job_name(self, provider):
+        """Test using job names on a simulator."""
         backend = provider.get_backend('ibmq_qasm_simulator')
         qobj = assemble(transpile(self._qc, backend=backend), backend=backend)
 
@@ -114,28 +90,8 @@ class TestIBMQJobAttributes(JobTestCase):
         self.assertEqual(job_id, retrieved_jobs[0].job_id())
 
     @requires_provider
-    def test_duplicate_job_name_backend(self, provider):
-        """Test multiple jobs with the same custom job name using a backend."""
-        backend = provider.get_backend('ibmq_qasm_simulator')
-        qobj = assemble(transpile(self._qc, backend=backend), backend=backend)
-
-        # Use a unique job name
-        job_name = str(time.time()).replace('.', '')
-        job_ids = set()
-        for _ in range(2):
-            job_ids.add(backend.run(qobj, job_name=job_name).job_id())
-
-        retrieved_jobs = backend.jobs(job_name=job_name)
-
-        self.assertEqual(len(retrieved_jobs), 2)
-        retrieved_job_ids = {job.job_id() for job in retrieved_jobs}
-        self.assertEqual(job_ids, retrieved_job_ids)
-        for job in retrieved_jobs:
-            self.assertEqual(job.name(), job_name)
-
-    @requires_provider
-    def test_duplicate_job_name_backend_service(self, provider):
-        """Test multiple jobs with the same custom job name using backend service."""
+    def test_duplicate_job_name(self, provider):
+        """Test multiple jobs with the same custom job name using a simulator."""
         backend = provider.get_backend('ibmq_qasm_simulator')
         qobj = assemble(transpile(self._qc, backend=backend), backend=backend)
 
@@ -147,6 +103,7 @@ class TestIBMQJobAttributes(JobTestCase):
 
         retrieved_jobs = provider.backends.jobs(backend_name=backend.name(),
                                                 job_name=job_name)
+
         self.assertEqual(len(retrieved_jobs), 2)
         retrieved_job_ids = {job.job_id() for job in retrieved_jobs}
         self.assertEqual(job_ids, retrieved_job_ids)
@@ -162,10 +119,15 @@ class TestIBMQJobAttributes(JobTestCase):
 
         job = backend.run(qobj)
         with self.assertRaises(IBMQJobFailureError):
-            job.result(timeout=300, partial=True)
+            job.result(timeout=300, partial=False)
 
         message = job.error_message()
         self.assertTrue(message)
+        self.assertIsNotNone(re.search(r'Error code: [0-9]{4}\.$', message), message)
+
+        r_message = provider.backends.retrieve_job(job.job_id()).error_message()
+        self.assertTrue(r_message)
+        self.assertIsNotNone(re.search(r'Error code: [0-9]{4}\.$', r_message), r_message)
 
     @requires_provider
     def test_error_message_simulator(self, provider):
@@ -183,6 +145,9 @@ class TestIBMQJobAttributes(JobTestCase):
         message = job.error_message()
         self.assertIn('Experiment 1: ERROR', message)
 
+        r_message = provider.backends.retrieve_job(job.job_id()).error_message()
+        self.assertIn('Experiment 1: ERROR', r_message)
+
     @requires_provider
     def test_error_message_validation(self, provider):
         """Test retrieving job error message for a validation error."""
@@ -194,6 +159,10 @@ class TestIBMQJobAttributes(JobTestCase):
 
         message = job.error_message()
         self.assertNotIn("Unknown", message)
+
+        # TODO Verify error code is in the message after API update
+        r_message = provider.backends.retrieve_job(job.job_id()).error_message()
+        self.assertEqual(message, r_message)
 
     @requires_provider
     def test_refresh(self, provider):